test(workflows): split runtime.test.ts along the runtime seam modules
The 1,381-line godfile becomes five seam-aligned files plus one shared
harness (SPEC ceiling: no test file over ~1,200 lines):

- runtime-engine.test.ts (584) — lib/engine.ts: createInstance,
  completeStep/advanceWorkflow, nested workflows (the createTask-node
  dispatch of lib/node-dispatch.ts is exercised inside these describes —
  no standalone node-dispatch describe existed to relocate)
- runtime-gates.test.ts (375) — lib/gates.ts: approve/reject, rewind,
  reopenFromStep, previousOutput on rejection
- runtime-step-context.test.ts (322) — lib/step-context.ts:
  getCurrentStep, skill resolution, deny_tools, getActiveAgents,
  agent-scoping/authorizeWorkflowToolUse
- runtime-instance-store.test.ts (133) — lib/instance-store.ts
- runtime-task-bridge.test.ts (132) — lib/task-bridge.ts (#482 seam)
- helpers/runtime-harness.ts — hook fakes, task-store/task-service module
  shapes, YAML fixtures + seed/reset; mock.module wiring stays per-file

Pure relocation: 73 pass / 188 expect() calls before and after. Each
split file keeps the dual content-dir mock and the dual task-store/
task-service specifiers. Other files' mock.module('…/lib/runtime')
overlays are unaffected — these tests exercise the real modules.
